Add a configure into the compilation routine
Replace the Makefile.config custom editing and guessing by a
'configure.sh' shell script that checks for dependencies, and generates
the appropriate configuration file.

This entails reworking the camlzip, extlib, src (Javalib) and toplevel
Makefiles to use the newly generated variables.

Local installation is now an easy option, and the use of
findlib/ocamlfind is generalized to all compilations and installations.

Signed-off-by: Florent Kirchner <florent.kirchner@gmail.com>