Use Stdlib instead of Pervasives (#22)
* Use Stdlib instead of Pervasives

Pervasives was deprecated in favor of directly using Stdlib in 4.08 and
then removed in the upcoming 5.0. This change makes javalib buildable
with the multi-core 5.0~alpha ocaml.

* Bump the package version to 3.2.2

NOTE: For some reason the package version in javalib.opam and in META
was 3.2 while in opam repo it's 3.2.1 (ditto for sawja dependency spec).
I changed the version to match x.y.z format.